requirements.txt
piexif==1.1.3
shortuuid==1.0.1
Pillow
pandas
smbprotocol